+++ apps/xmtp.chat/src/components/App/App.tsx
@@ -11,7 +11,7 @@ import { ManageMembersModal } from "@/components/Conversation/ManageMembersModal
 import { ManagePermissionsModal } from "@/components/Conversation/ManagePermissionsModal";
 import { ManagePropertiesModal } from "@/components/Conversation/ManagePropertiesModal";
 import { NewConversation } from "@/components/Conversation/NewConversation";
-import { Identity } from "@/components/Identity/Identity";
+import { IdentityModal } from "@/components/Identity/IdentityModal";
 import { MessageModal } from "@/components/Messages/MessageModal";
 import { useAnalytics } from "@/hooks/useAnalytics";
 
@@ -27,12 +27,13 @@ export const App: React.FC = () => {
         </Route>
         <Route path="/*" element={<AppLayout />}>
           <Route index element={<Navigate to="/conversations" />} />
-          <Route path="identity" element={<Identity />} />
           <Route path="dm/:address" element={<LoadDM />} />
           <Route path="conversations">
             <Route index element={<SelectConversation />} />
+            <Route path="identity" element={<IdentityModal />} />
             <Route path="new" element={<NewConversation />} />
             <Route path=":conversationId" element={<LoadConversation />}>
+              <Route path="identity" element={<IdentityModal />} />
               <Route path="message/:messageId" element={<MessageModal />} />
               <Route path="manage">
                 <Route path="consent" element={<ManageConsentModal />} />
